Add a warning about the Aniseed client eventually being deprecated
This is in the hopes that we can get people to pin to the Aniseed client
if they care about this, ignore it if they don't care and be aware of
the change if they're looking forward to it.

If this is just annoying and unhelpful maybe I'll remove it, I'd just
rather start to spread awareness.